Fixes problems for adding event, location and product

+++ src/main/java/ch/roester/event/EventRequestDTO.java
@@ -14,7 +14,7 @@ public class EventRequestDTO {
     private String description;
 
     @NotNull
-    private Long locationId;
+    private Integer locationId;
 
     @NotNull
     private LocalDateTime date;
